core: ardupilot-manager: Move type definitions to typedefs file
