+++ core/services/ardupilot_manager/firmware/test_FirmwareDownload.py
@@ -2,7 +2,8 @@ import os
 
 import pytest
 
-from firmware.FirmwareDownload import FirmwareDownloader, Platform, Vehicle
+from firmware.FirmwareDownload import FirmwareDownloader
+from typedefs import Platform, Vehicle
 
 
 def test_static() -> None:
